User authentication

See merge request juniors24apr/anthill-v2!38

+++ actions/tracks/delete-track.ts
@@ -1,11 +1,16 @@
 "use server"
 
 import prisma from "@/lib/db"
-import { handleError } from "@/utils/error-handler"
+import { CustomError, handleError } from "@/utils/error-handler"
 import tracksPrismaCodesMap from "@/utils/prisma-codes/tracks-prisma-codes"
+import { auth } from "@clerk/nextjs/server"
 
 const deleteTrack = async (trackId: string) => {
   try {
+    const { userId } = auth()
+    if (!userId) {
+      throw new CustomError("User not authenticated", "NOT_AUTHENTICATED")
+    }
     return await prisma.track.delete({ where: { id: trackId } })
   } catch (error) {
     return handleError(error, tracksPrismaCodesMap)
